Added `ser/serializable` module
